MyChalls/Chall1/requirements.txt
flask
Pillow
playwright